+++ index.html
@@ -8,7 +8,6 @@
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
 
   <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
 
   <title>Кекстаграм</title>
 </head>
@@ -236,5 +235,6 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
