+++ js/main.js
@@ -1,63 +1,3 @@
-const NAMES = [
-  'Артём',
-  'Иван',
-  'Ольга',
-  'Алина',
-  'Дима',
-  'Мария',
-  'Алексей',
-  'Анна',
-  'Николай',
-  'Татьяна'
-];
+import { generatePhotos } from './data.js';
 
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const Like = {
-  MIN: 15,
-  MAX: 200
-};
-
-const Comment = {
-  MIN: 1,
-  MAX: 10
-};
-
-const IdComment = {
-  MIN: 1,
-  MAX: 100
-};
-
-const PHOTO_NUMBER = 25;
-
-const getRandomNumber = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
-
-const getRandomElement = (items) => items[getRandomNumber(0, items.length - 1)];
-
-const generateComment = () => ({
-  id: getRandomNumber(IdComment.MIN, IdComment.MAX),
-  avatar: `img/avatar-${getRandomNumber(1, 6)}.svg`,
-  message: getRandomElement(MESSAGES),
-  name: getRandomElement(NAMES)
-});
-
-const generateComments = () => Array.from({ length: getRandomNumber(Comment.MIN, Comment.MAX)}, (v, i) => generateComment(i + 1));
-
-const generatePhoto = (id) => ({
-  id: id,
-  url: `photos/${id}.jpg`,
-  description: `Описание фотографии номер ${id}` ,
-  likes: getRandomNumber(Like.MIN, Like.MAX),
-  comments: generateComments()
-});
-
-const generatePhotos = () => Array.from({ length: PHOTO_NUMBER }, (v, i) => generatePhoto(i + 1));
-
-generatePhotos();
+generatePhotos ();
